requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
anyio==4.12.1
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
datasets==4.7.0
dill==0.4.0
filelock==3.25.2
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.1
idna==3.11
joblib==1.5.3
kiwisolver==1.5.0
markdown-it-py==4.0.0
matplotlib==3.10.8
mdurl==0.1.2
multidict==6.7.1
multiprocess==0.70.18
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pillow==12.1.1
propcache==0.4.1
pyarrow==23.0.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.32.5
rich==14.3.3
scikit-learn==1.8.0
scipy==1.17.1
shellingham==1.5.4
six==1.17.0
threadpoolctl==3.6.0
tqdm==4.67.3
typer==0.24.1
typing_extensions==4.15.0
urllib3==2.6.3
xxhash==3.6.0
yarl==1.23.0
